Migrate Mutation.enableExternalSchemaComposition
